run with bun instead of node
bun doesn't polyfill Bun when bundled to run in nodejs

there are plans for bun to polyfill in the future.
when they do, we can switch this back to run with nodejs
instead of bun for better compatibility and so people
don't need to install bun to run the plugin.

I spent a while migrating this project to bun without
realizing that was a limitation and I'm not going back at this moment