fix(get-entry): fix unloaded buffer response
the list_buffers function returns unloaded/delete buffers. sometimes we would get a match for a file
and try to get its contents. the function to get lines when called with an unloaded buffer id,
returns an empty array which is what we were returning to the browser. we now check if a buffer is
loaded/valid before getting its contents

fix #154