fix(settings): don't close settings on color pick
what triggers the timeout to close the settings modal is when the cursor leaves the modal, but when
picking a color, the mouse would need to leave the modal to interact with the color picker. the
mouse leaving the modal would cause the modal to close. this fixes that and prevents the timeout
timer from starting if the user if the color picker is open

fix #191