fix(single-file): don't search for root if single-file mode is enabled
Even if single-file mode was enabled, we were still trying to find a git repository root. In some cases, depending on the cwd of the user, the search could take a while and it would seem like the app froze. Now if the user explicitly enabled single-file mode, we don't even try searching for a repository root.

re #249